rename `generate_insertion_fasta` to `insertions_to_fasta`
