+++ src/DAJIN2/core/preprocess/insertions_to_fasta.py
@@ -5,12 +5,14 @@ from collections import defaultdict
 from itertools import groupby
 from pathlib import Path
 
-import midsv
+
 import numpy as np
 from rapidfuzz import process
 from rapidfuzz.distance import DamerauLevenshtein
 from sklearn import metrics
 from sklearn.cluster import MeanShift, MiniBatchKMeans
+
+from DAJIN2.utils import io
 
 ###########################################################
 # Detect insertion sequences
@@ -19,34 +21,35 @@ from sklearn.cluster import MeanShift, MiniBatchKMeans
 
 def _count_insertions(path: Path | str, mutation_loci: dict) -> dict[tuple[int, str], int]:
     insertion_counts = defaultdict(int)
-    for m in midsv.read_jsonl(path):
+    for m in io.read_jsonl(path):
         cssplits = m["CSSPLIT"].split(",")
         for idx in (i for i, m in enumerate(mutation_loci) if "+" in m):
             if cssplits[idx].startswith("+"):
                 insertion_counts[(idx, cssplits[idx])] += 1
-    # remove low frequency insertions
-    coverage_sample = sum(1 for _ in midsv.read_jsonl(path))
+
+    # Remove low frequency insertions
+    coverage_sample = sum(1 for _ in io.read_jsonl(path))
     threshold = int(coverage_sample * 0.5 / 100)
-    insertion_counts = {k: v for k, v in insertion_counts.items() if v >= threshold}
-    return insertion_counts
+    return {k: v for k, v in insertion_counts.items() if v >= threshold}
 
 
 def _create_insertions_dict(
     insertion_sample: dict[tuple[int, str], int], insertion_control: dict[tuple[int, str], int]
-) -> defaultdict[dict[str, int]]:
+) -> defaultdict[int, dict[str, int]]:
+    """Create a dictionary of insertions that are present in the sample but not in the control."""
     insertions = defaultdict(dict)
     for key in insertion_sample.keys() - insertion_control.keys():
         score = insertion_sample[key]
         idx, seq = key
         if isinstance(seq, int):
             idx, seq = seq, idx
-        insertions[idx].update({seq: score})
+        insertions[idx][seq] = score
     return insertions
 
 
 def extract_insertions(
     path_sample: Path | str, path_control: Path | str, mutation_loci: dict
-) -> defaultdict[dict[str, int]]:
+) -> defaultdict[int, dict[str, int]]:
     insertion_sample = _count_insertions(path_sample, mutation_loci)
     insertion_control = _count_insertions(path_control, mutation_loci)
     return _create_insertions_dict(insertion_sample, insertion_control)
@@ -145,7 +148,7 @@ def merge_similar_insertions(insertions, mutation_loci) -> dict[dict[frozenset[s
 def extract_score_and_sequence(path_sample, insertions_merged) -> list[tuple[list[int], str]]:
     scores = []
     sequences = []
-    for m in midsv.read_jsonl(path_sample):
+    for m in io.read_jsonl(path_sample):
         score = defaultdict(int)
         seq = defaultdict(lambda: "N")
         cssplits = m["CSSPLIT"].split(",")
